Moved a bunch of code around to dry it out
