+++ src/client.rs
@@ -11,14 +11,10 @@ use self::subcomponents::{
 use crate::ResponseResult;
 use serde::de::DeserializeOwned;
 use std::future::Future;
-use std::ops::RangeFrom;
 
 /// A `Client` is used to make multiple requests to a specific zcashd RPC server. Requests are invoked by async methods that correspond to `zcashd` RPC API method names with request-specific parameters. Each such method has an associated response type.
 pub struct Client {
-    url: String,
-    auth: String,
-    reqcli: reqwest::Client,
-    idit: RangeFrom<u64>,
+    inner: utils::InnerCli,
 }
 
 impl Client {
@@ -27,10 +23,7 @@ impl Client {
     /// - `authcookie` is the contents of `~/.zcash/.cookie`.
     pub fn new(hostport: String, authcookie: String) -> Client {
         Client {
-            url: format!("http://{}/", hostport),
-            auth: format!("Basic {}", base64::encode(authcookie)),
-            reqcli: reqwest::Client::new(),
-            idit: (0..),
+            inner: utils::InnerCli::new(hostport, authcookie),
         }
     }
 
@@ -51,18 +44,9 @@ impl Client {
     where
         R: DeserializeOwned,
     {
-        use crate::{
-            envelope::{RequestEnvelope, ResponseEnvelope},
-            json,
-        };
+        use crate::{envelope::ResponseEnvelope, json};
 
-        let id = self.idit.next().unwrap();
-        let sendfut = self
-            .reqcli
-            .post(&self.url)
-            .header("Authorization", &self.auth)
-            .body(&RequestEnvelope::wrap(id, method, args))
-            .send();
+        let (id, sendfut) = self.inner.procedure_call(method, args);
         async move {
             let reqresp = sendfut.await?;
             let text = reqresp.text().await?;
