+++ src/client/utils.rs
@@ -37,3 +37,68 @@ pub fn get_zcashd_port() -> String {
 pub fn make_client(regtest: bool) -> crate::Client {
     crate::Client::new(get_zcashd_port(), get_cookie(regtest).unwrap())
 }
+
+#[derive(Debug, serde::Deserialize, serde::Serialize)]
+pub struct RequestEnvelope {
+    pub(crate) id: u64,
+    pub(crate) method: &'static str,
+    pub(crate) params: Vec<serde_json::Value>,
+}
+
+impl<'a> From<&'a RequestEnvelope> for reqwest::Body {
+    fn from(re: &'a RequestEnvelope) -> reqwest::Body {
+        use serde_json::to_string_pretty;
+
+        reqwest::Body::from(to_string_pretty(re).unwrap())
+    }
+}
+
+impl RequestEnvelope {
+    pub fn wrap(
+        id: u64,
+        method: &'static str,
+        params: Vec<serde_json::Value>,
+    ) -> RequestEnvelope {
+        RequestEnvelope {
+            id: id,
+            method: method,
+            params: params,
+        }
+    }
+}
+
+pub(crate) struct InnerCli {
+    pub(crate) url: String,
+    pub(crate) auth: String,
+    pub(crate) reqcli: reqwest::Client,
+    pub(crate) idit: std::ops::RangeFrom<u64>,
+}
+
+impl InnerCli {
+    pub(crate) fn new(hostport: String, authcookie: String) -> Self {
+        Self {
+            url: format!("http://{}/", hostport),
+            auth: format!("Basic {}", base64::encode(authcookie)),
+            reqcli: reqwest::Client::new(),
+            idit: (0..),
+        }
+    }
+    pub(crate) fn procedure_call(
+        &mut self,
+        method: &'static str,
+        args: Vec<serde_json::Value>,
+    ) -> (
+        u64,
+        impl std::future::Future<Output = Result<reqwest::Response, reqwest::Error>>,
+    ) {
+        let id = self.idit.next().unwrap();
+        (
+            id,
+            self.reqcli
+                .post(&self.url)
+                .header("Authorization", &self.auth)
+                .body(&RequestEnvelope::wrap(id, method, args))
+                .send(),
+        )
+    }
+}
